+++ Kms Cloud Database/UserDataHourlyDistance.cs
@@ -18,9 +18,9 @@ namespace Kms.Cloud.Database
         public System.DateTime Timestamp { get; set; }
         public long Steps { get; set; }
         public long Distance { get; set; }
-        public long Kcal { get; set; }
-        public long Co2 { get; set; }
-        public long Cash { get; set; }
+        public double Kcal { get; set; }
+        public double Co2 { get; set; }
+        public double Cash { get; set; }
     
         public virtual User User { get; set; }
     }
